Refractor code to use more OOP
