--- data/duke.txt
@@ -1,4 +1,0 @@
-T| |borrow book
-T| |buy milk
-D| |return book|2022-03-05
-E| |project meeting|2022-04-26|12:00|14:30
