--- text-ui-test/data/duke.txt
@@ -1,3 +1,0 @@
-T| |buy milk
-D| |return book|Sunday
-E| |project meeting|Mon 2-4pm
